Bugfix navigation simulation distance increasing and bearing facing the wrong direction
Edge case where the user starts on the route ahead of the first checkpoint. As a result, the user "does not pass" the first checkpoint, so it does not trigger the loading of the next checkpoint. This makes the user always face the first checkpoint, and move away from it causing the issues observed.